// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/company-details-model.ts
export class CompanyDetailsModel {
    companyId: string;
    companyName: string;
    siteAddress: string;
    workEmail: string;
    password: string;
    industryId: string;
    mainUseCaseId: string;
    phoneNumber: string;
    countryId: string;
    timeZoneId: string;
    currencyId: string;
    numberFormatId: string;
    dateFormatId: string;
    timeFormatId: string;
    teamSize: number;
    versionNumber: string;
    inActiveDateTime: Date;
    originalId: string;
    createdDateTime: Date;
    createdByUserId: string;
}